+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Note that since we don't clearly distinguish between a public and private interf
   - Primitive arrow - nicer default cap size (relative to tube_radius)
   - Primitive angle_measurement - added vector_radius param
   - Fix MVSX file assets being disposed in multi-snapshot states
-- Show FOV adjusted position in `CameraInfo` UI and use it in "Copy MVS State"
+- Add `mol-utils/camera.ts` with `fovAdjustedPosition` and `fovNormalizedCameraPosition`
+- Show FOV normalized position in `CameraInfo` UI and use it in "Copy MVS State"
 - Support static resources in `AssetManager`